get markers with get request. Add defaultValue for newMarker
