+++ src/components/features/actionsMenu/DeliveriesTableActionsMenu.tsx
@@ -3,10 +3,10 @@ import Menu from '@mui/material/Menu'
 import MenuItem from '@mui/material/MenuItem'
 import MoreHorizIcon from '@mui/icons-material/MoreHoriz'
 import React from 'react'
-import WarningActionDialog from '../shared/WarningActionDialog'
 import { useTranslation } from 'react-i18next'
+import WarningActionDialog from '@/components/shared/WarningActionDialog'
 
-export default function VendorTableActionsMenu() {
+export default function DeliveriesTableActionsMenu() {
   const { t: translate } = useTranslation()
   const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null)
   const open = Boolean(anchorEl)
@@ -32,7 +32,11 @@ export default function VendorTableActionsMenu() {
     setSelectedOption(option)
   }
 
-  const options = [translate('actionsMenu.options.edit'), translate('actionsMenu.options.delete')]
+  const options = [
+    translate('actionsMenu.options.details'),
+    translate('actionsMenu.options.approve'),
+    translate('actionsMenu.options.delete')
+  ]
 
   return (
     <div>
@@ -60,13 +64,14 @@ export default function VendorTableActionsMenu() {
         ))}
       </Menu>
 
-      {selectedOption === 'Изтрий' && (
+      {/* TODO: Only admin action */}
+      {selectedOption === translate('actionsMenu.options.delete') && (
         <WarningActionDialog
           open={open}
-          title={translate('deleteAction.vendors.title')}
-          content={translate('deleteAction.vendors.message')}
-          discardText={translate('deleteAction.vendors.labels.discard')}
-          confirmText={translate('deleteAction.vendors.labels.confirm')}
+          title={translate('deliveries.deleteActions.title')}
+          content={translate('deliveries.deleteActions.message')}
+          discardText={translate('deliveries.deleteActions.labels.discard')}
+          confirmText={translate('deliveries.deleteActions.labels.confirm')}
           onCloseDialog={handleClose}
           onDiscardClick={onDiscardClick}
           onConfirmClick={onConfirmClick}
