+++ src/components/features/actionsMenu/ZonesTableActionsMenu.tsx
@@ -3,8 +3,8 @@ import Menu from '@mui/material/Menu'
 import MenuItem from '@mui/material/MenuItem'
 import MoreHorizIcon from '@mui/icons-material/MoreHoriz'
 import React from 'react'
-import WarningActionDialog from '../shared/WarningActionDialog'
 import { useTranslation } from 'react-i18next'
+import WarningActionDialog from '@/components/shared/WarningActionDialog'
 
 export default function ZonesTableActionsMenu() {
   const { t: translate } = useTranslation()
