+++ src/components/features/admin/ZonesTable.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import DataTable from '@/components/shared/DataTable'
 import { useTranslation } from 'react-i18next'
-import ZonesTableActionsMenu from '../ZonesTableActionsMenu'
 import ChipsList from '../ChipsList'
 import SearchInput from '../SearchInput'
 import { Autocomplete, TextField } from '@mui/material'
-import { Column } from '@/interfaces/column.ts'
+import { Column } from '@/interfaces/Column.ts'
+import ZonesTableActionsMenu from '../actionsMenu/ZonesTableActionsMenu'
 
 interface Row {
   name: string
