+++ src/components/shared/TableActionsMenu.tsx
@@ -1,83 +1,82 @@
-import IconButton from '@mui/material/IconButton'
-import Menu from '@mui/material/Menu'
-import MenuItem from '@mui/material/MenuItem'
-import MoreHorizIcon from '@mui/icons-material/MoreHoriz'
-import React from 'react'
-import WarningActionDialog from '../shared/WarningActionDialog'
-import { useTranslation } from 'react-i18next'
+// import IconButton from '@mui/material/IconButton'
+// import Menu from '@mui/material/Menu'
+// import MenuItem from '@mui/material/MenuItem'
+// import MoreHorizIcon from '@mui/icons-material/MoreHoriz'
+// import React from 'react'
+// import WarningActionDialog from '../shared/WarningActionDialog'
+// import { useTranslation } from 'react-i18next'
 
-interface TableActionsMenuProps {
-  itemProps: string[];
-  page: string;
-}
+// interface TableActionsMenuProps {
+//   itemProps: string[];
+//   page: string;
+// }
 
-export default function TableActionsMenu({ itemProps, page }: TableActionsMenuProps) {
-  const { t: translate } = useTranslation()
-  const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null)
-  const open = Boolean(anchorEl)
-  const [selectedOption, setSelectedOption] = React.useState<string | null>(null)
+// export default function TableActionsMenu({ itemProps, page }: TableActionsMenuProps) {
+//   const { t: translate } = useTranslation()
+//   const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null)
+//   const open = Boolean(anchorEl)
+//   const [selectedOption, setSelectedOption] = React.useState<string | null>(null)
 
-  const handleClick = (event: React.MouseEvent<HTMLElement>) => {
-    setAnchorEl(event.currentTarget)
-  }
-  const handleClose = () => {
-    setSelectedOption(null)
-    setAnchorEl(null)
-  }
+//   const handleClick = (event: React.MouseEvent<HTMLElement>) => {
+//     setAnchorEl(event.currentTarget)
+//   }
+//   const handleClose = () => {
+//     setSelectedOption(null)
+//     setAnchorEl(null)
+//   }
 
-  const onDiscardClick = () => {
-    handleClose()
-  }
+//   const onDiscardClick = () => {
+//     handleClose()
+//   }
 
-  const onConfirmClick = () => {
-    handleClose()
-  }
+//   const onConfirmClick = () => {
+//     handleClose()
+//   }
 
-  const actionHandler = (option: string) => {
-    setSelectedOption(option)
-  }
+//   const actionHandler = (option: string) => {
+//     setSelectedOption(option)
+//   }
 
-  const options = itemProps.map((str) => translate(`actionsMenu.options.${str}`))
+//   const options = itemProps.map((str) => translate(`actionsMenu.options.${str}`))
 
+//   return (
+//     <div>
+//       <IconButton
+//         aria-label="more"
+//         id="long-button"
+//         aria-controls={open ? 'long-menu' : undefined}
+//         aria-expanded={open ? 'true' : undefined}
+//         aria-haspopup="true"
+//         onClick={handleClick}>
+//         <MoreHorizIcon />
+//       </IconButton>
+//       <Menu
+//         id="long-menu"
+//         MenuListProps={{
+//           'aria-labelledby': 'long-button'
+//         }}
+//         anchorEl={anchorEl}
+//         open={open}
+//         onClose={handleClose}>
+//         {options.map((option) => (
+//           <MenuItem key={option} onClick={() => actionHandler(option)}>
+//             {option}
+//           </MenuItem>
+//         ))}
+//       </Menu>
 
-  return (
-    <div>
-      <IconButton
-        aria-label="more"
-        id="long-button"
-        aria-controls={open ? 'long-menu' : undefined}
-        aria-expanded={open ? 'true' : undefined}
-        aria-haspopup="true"
-        onClick={handleClick}>
-        <MoreHorizIcon />
-      </IconButton>
-      <Menu
-        id="long-menu"
-        MenuListProps={{
-          'aria-labelledby': 'long-button'
-        }}
-        anchorEl={anchorEl}
-        open={open}
-        onClose={handleClose}>
-        {options.map((option) => (
-          <MenuItem key={option} onClick={() => actionHandler(option)}>
-            {option}
-          </MenuItem>
-        ))}
-      </Menu>
-
-      {selectedOption === 'Изтрий' && (
-        <WarningActionDialog
-          open={open}
-          title={translate(`${page}.deleteAction.title`)}
-          content={translate(`${page}.deleteAction.message`)}
-          discardText={translate(`${page}.deleteAction.labels.discard`)}
-          confirmText={translate(`${page}.deleteAction.labels.confirm`)}
-          onCloseDialog={handleClose}
-          onDiscardClick={onDiscardClick}
-          onConfirmClick={onConfirmClick}
-        />
-      )}
-    </div>
-  )
-}
+//       {selectedOption === 'Изтрий' && (
+//         <WarningActionDialog
+//           open={open}
+//           title={translate(`${page}.deleteAction.title`)}
+//           content={translate(`${page}.deleteAction.message`)}
+//           discardText={translate(`${page}.deleteAction.labels.discard`)}
+//           confirmText={translate(`${page}.deleteAction.labels.confirm`)}
+//           onCloseDialog={handleClose}
+//           onDiscardClick={onDiscardClick}
+//           onConfirmClick={onConfirmClick}
+//         />
+//       )}
+//     </div>
+//   )
+// }
